refactor: split NeatConfigParsers.ts by concern (#2396)
Split the 22 parser functions in src/config/NeatConfigParsers.ts (897
lines) into seven focused sub-modules under src/config/parsers/, grouped
by concern:

- RuntimeParsers: worker thread cap, WASM cache, memory, parallel eval
- DiscoveryParsers: discovery min candidates, cache, disk space
- MutationParsers: adaptive mutation, plateau detection, stability, MCMC
- PopulationParsers: fine-tune, adaptive population, ensemble diversity
- TrainingParsers: hyperparameter, quantum step, predictive coding, CV
- RegularisationParsers: weight, bias regularisation
- DataParsers: data fuzzing, data quantisation

NeatConfigParsers.ts is now a 45-line barrel re-export so existing import
paths (@config/NeatConfigParsers.ts) continue to work for tests and
NeatConfig.ts itself. Every new module is well under the 250-line cap
(largest is MutationParsers at 205 lines).

Added 69 unit tests under test/config/parsers/ covering happy-path
defaults, override application, and invalid-input rejection for every
parser. No behavioural change — full quality gate green (6118 tests
pass).
